+++ CHANGELOG.md
@@ -14,3 +14,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `CultureInfo` extension method to get the language code of a culture
 - `IEnumerable` extension methods
 - `IFormFile` extension method (AspNetCore package only)
+- `DbSet` extension methods for ISortableEntity interface (EntityFrameworkCore package only)